backend/requirements.txt
# Core FastAPI & WebSocket
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-dotenv==1.0.1
python-socketio==5.11.0
aiohttp==3.9.1

# LangChain & AI (v1.0+)
langchain>=1.0.0
langchain-openai>=0.2.5
langchain-community>=0.3.5
langgraph-prebuilt==1.0.2
openai>=1.54.3

# Pydantic (for data models)
pydantic>=2.9.2
pydantic-settings>=2.6.0

# Additional utilities
python-multipart==0.0.12
aiofiles==24.1.0
